refactor: move TrackInfo and addPointsToTrack to new "logic" folder
