+++ logic/TrackInfo.js
@@ -1,0 +1,11 @@
+
+class TrackInfo {
+  constructor(track,trackData)
+  {
+    this.track = track;
+    this.trackData = trackData;
+  }
+}
+
+
+module.exports = TrackInfo
