+++ routes/api/tracks.js
@@ -6,132 +6,8 @@ var Comment = mongoose.model('Comment');
 var User = mongoose.model('User');
 var auth = require('../auth');
 var currentTracks = new Map();
-class trackInfo {
-  constructor (track,trackData)
-  {
-    this.track = track;
-    this.trackData = trackData;
-  }
-}
-
-
-function addPointsToTrack(track,body)
-{
-  var num = 0;
-  var start = 0;
-  var end = 0;
- //console.log("len"+body.length);
-  while (end < body.length) {
-    start = end;
-    while (body[end] != ";" && body[end] != "$" && end < body.length) {
-      end++;
-    }
-    if(body[end] == "$") // $ is replacing \n as newlines are not allowed in json strings
-    {
-        num=0;
-    }
-    if(end < body.length)
-    {
-    var token = body.substr(start, end - start);
-    end++;
-    if(token.length>0)
-    {
-    //console.log(token);
-    //console.log("num:"+num);
-    //console.log("end:"+end);
-    
-    if((num == 0) && (token == "Date"))
-    {
-      // we have a header line, ignore it for now, TODO parse it
-      if (end < body.length) {
-        while (body[end] != ";" && body[end] != "$" && end < body.length) {
-          end++;
-        }
-        start = end;
-        num=100;
-      }
-    }
-    if (num == 0) {
-      track.trackData.points.push({ date: "dummy", time: "", latitude: "", longitude: "", course: "", speed: "", d1: "", d2: "", flag: "", private: ""});
-      track.trackData.points[track.trackData.points.length - 1].date = token;
-      num++;
-    }
-    else if (num == 1) {
-      track.trackData.points[track.trackData.points.length - 1].time = token;
-      num++;
-    }
-    else if (num == 2) {
-      var f = parseFloat(token);
-      if(isNaN(f))
-      {
-          f = parseFloat(token.substring(0,10));
-      }
-      if(isNaN(f))
-      {
-        f=0.0;
-      }
-      track.trackData.points[track.trackData.points.length - 1].latitude = f;
-      num++;
-    }
-    else if (num == 3) {
-      var f = parseFloat(token);
-      if(isNaN(f))
-      {
-          f = parseFloat(token.substring(0,10));
-      }
-      if(isNaN(f))
-      {
-        f=0.0;
-      }
-      track.trackData.points[track.trackData.points.length - 1].longitude = f;
-      num++;
-    }
-    else if (num == 4) {
-      var f = parseFloat(token);
-      if(isNaN(f))
-      {
-          f = parseFloat(token.substring(0,10));
-      }
-      if(isNaN(f))
-      {
-        f=0.0;
-      }
-      track.trackData.points[track.trackData.points.length - 1].course = f;
-      num++;
-    }
-    else if (num == 5) {
-      var f = parseFloat(token);
-      if(isNaN(f))
-      {
-          f = parseFloat(token.substring(0,10));
-      }
-      if(isNaN(f))
-      {
-        f=0.0;
-      }
-      track.trackData.points[track.trackData.points.length - 1].speed = f;
-      num++;
-    }
-    else if (num == 6) {
-      track.trackData.points[track.trackData.points.length - 1].d1 = token;
-      num++;
-    }
-    else if (num == 7) {
-      track.trackData.points[track.trackData.points.length - 1].d2 = token;
-      num++;
-    }
-    else if (num == 8) {
-      track.trackData.points[track.trackData.points.length - 1].flag = token;
-      num++;
-    }
-    else if (num == 9) {
-      track.trackData.points[track.trackData.points.length - 1].private = token;
-      num++;
-    }
-    }
-    }
-  }
-}
+var TrackInfo = require('../../logic/TrackInfo');
+var {addPointsToTrack} = require('../../logic/tracks');
 
 // Preload track objects on routes with ':track'
 router.param('track', function(req, res, next, slug) {
@@ -291,7 +167,7 @@ router.post('/', auth.required, function(req, res, next) {
     var track = new Track(req.body.track);
     var trackData = new TrackData();
     track.trackData = trackData._id;
-     
+
     track.author = user;
     track.visible = track.author.areTracksVisibleForAll;
     trackData.save(function (err){
@@ -318,7 +194,7 @@ router.post('/add', auth.optional, function(req, res, next) {
     var ti = null;
     if (currentTracks.has(req.body.id))
       ti = currentTracks.get(req.body.id);
-      
+
     //console.log("TI" + ti);
     //console.log("TILen" + ti.trackData.points.length);
     //console.log("TITrack" + ti.track);
@@ -345,12 +221,12 @@ router.post('/begin', auth.optional, function (req, res, next) {
 
     if(currentTracks.has(req.body.id))
         currentTracks.delete(req.body.id); // delete old parts if there are leftovers
-    var ti = new trackInfo(new Track(req.body.track),new TrackData());
-    ti.track.trackData = ti.trackData._id; 
+    var ti = new TrackInfo(new Track(req.body.track),new TrackData());
+    ti.track.trackData = ti.trackData._id;
     currentTracks.set(req.body.id, ti);
 
   //console.log("addToTrack"+req.body);
-    
+
     addPointsToTrack(ti, ti.track.body);
 
       //console.log("TLen" + ti.track);
@@ -381,7 +257,7 @@ router.post('/end', auth.optional, function (req, res, next) {
     }
     else
     {
-      var ti = new trackInfo(new Track(req.body.track),new TrackData());
+      var ti = new TrackInfo(new Track(req.body.track),new TrackData());
       ti.track.trackData = ti.trackData._id;
       addPointsToTrack(ti, ti.track.body);
     }
